03 Refactor with Clean Architecture
Rename every mention to `%ApiClient` to `Repository`.

In the future these files will work with an Entity Framework Core database.
