+++ src/BreweryApi.Application/Services/BreweryService.cs
@@ -5,33 +5,33 @@ namespace BreweryApi.Application.Services;
 
 public class BreweryService : IBreweryService
 {
-    private IApiClient _breweryClient;
-    public BreweryService(IApiClient breweryClient)
+    private IBreweryRepository _breweryRepository;
+    public BreweryService(IBreweryRepository breweryRepository)
     {
-        _breweryClient = breweryClient;
+        _breweryRepository = breweryRepository;
     }
     public IEnumerable<Brewery> GetAllBreweries()
     {
-        return _breweryClient.GetAllBreweries();
+        return _breweryRepository.GetAllBreweries();
     }
 
     public Brewery  GetBreweryById(Guid breweryId)
     {
-        return _breweryClient.GetBreweryById(breweryId);
+        return _breweryRepository.GetBreweryById(breweryId);
     }
 
     public Brewery  CreateBrewery(Brewery brewery)
     {
-        return _breweryClient.CreateBrewery(brewery);
+        return _breweryRepository.CreateBrewery(brewery);
     }
 
     public Brewery UpdateBrewery(Guid id, Brewery brewery)
     {
-        return _breweryClient.UpdateBrewery(id, brewery);
+        return _breweryRepository.UpdateBrewery(id, brewery);
     }
 
     public IEnumerable<Brewery> DeleteBrewery(Guid id)
     {
-        return _breweryClient.DeleteBrewery(id);
+        return _breweryRepository.DeleteBrewery(id);
     }
 }
