+++ src/BreweryApi.Infrastructure/Extensions/DependencyInjection.cs
@@ -1,5 +1,5 @@
 using BreweryApi.Application.Abstractions;
-using BreweryApi.Infrastructure.Clients;
+using BreweryApi.Infrastructure.Repositories;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace BreweryApi.Infrastructure.Extensions;
@@ -8,7 +8,7 @@ public static class DependencyInjection
 {
     public static IServiceCollection RegisterServices(this IServiceCollection services)
     {
-        services.AddSingleton<IApiClient, BreweryApiClient>();
+        services.AddSingleton<IBreweryRepository, BreweryRepository>();
         return services;
     }
 }
